+++ src/directives/ngie-ignore.js
@@ -9,14 +9,8 @@ function $$ngieIgnoreFactory() {
         priority: 1,
         restrict: 'AECM',
         link($s, el, attrs) {
-
-            console.log('IN DIRECTIVE');
-
-
-            // TODO fix matching here and in $compile
-            console.log('HTML', el.html().replace(/(\{{2,3}[^\}]+\}{2,3})/g, '{{{\'$1\'}}}'));
-
             el.html(el.html().replace(/(\{{2,3}[^\}]+\}{2,3})/g, '{{{\'$1\'}}}'));
+
             el.removeAttr('ngie-ignore');
             delete attrs.ngieIgnore;
         }