+++ test/src/Server.spec.js
@@ -11,6 +11,7 @@ import $LogProvider from            'angie-log';
 // Angie Modules
 import app from                     '../../src/Angie';
 import { $$server } from            '../../src/Server';
+import { config } from              '../../src/Config';
 import $Request from                '../../src/services/$Request';
 import * as $Responses from         '../../src/services/$Response';
 
@@ -42,7 +43,8 @@ describe('$$server', function() {
         };
         response = {
             end,
-            _header: 'test'
+            _header: 'test',
+            setHeader: spy()
         };
         mock(app, '$$load', function() {
             return {
@@ -112,6 +114,7 @@ describe('$$server', function() {
     it('test call with http', function() {
         $$server([ 'server', 1234 ]);
         assert(app.$$load.called);
+        expect(response.setHeader.callCount).to.eq(3);
         assert(http.createServer.called);
         expect(https.createServer).to.not.have.been.called;
         assert(global.setTimeout.called);
@@ -134,10 +137,39 @@ describe('$$server', function() {
         expect(end.callCount).to.eq(2);
         expect($LogProvider.info.calls[0].args[0]).to.eq('Serving on port 1234');
     });
+    it('test call with http, X-Frame-Options config set', function() {
+        config.setXFrameOptions = 'SAMEORIGIN';
+        $$server([ 'server', 1234 ]);
+        assert(app.$$load.called);
+        expect(response.setHeader.callCount).to.eq(4);
+        assert(http.createServer.called);
+        expect(https.createServer).to.not.have.been.called;
+        assert(global.setTimeout.called);
+        assert(global.clearTimeout.called);
+        assert(dataMock.called);
+        assert(routeMock.called);
+        expect($Responses.ErrorResponse.calls[0].args[0]).to.deep.eq(e);
+        expect(
+            $LogProvider.error.calls[0].args
+        ).to.deep.eq([ 'GET', 'test', 'test' ]);
+        assert(head.called);
+        assert(writeSync.called);
+        expect(listen.calls[0].args[0]).to.eq(1234);
+        expect(
+            app.$$tearDown.calls[0].args
+        ).to.deep.eq([ '$request', '$response' ]);
+        expect(
+            app.$$tearDown.calls[1].args
+        ).to.deep.eq([ '$request', '$response' ]);
+        expect(end.callCount).to.eq(2);
+        expect($LogProvider.info.calls[0].args[0]).to.eq('Serving on port 1234');
+        delete config.setXFrameOptions;
+    });
     it('test call with http and argument -p', function() {
         yargs([ '-p', '9999' ]);
         $$server([ 'server', 1234 ]);
         assert(app.$$load.called);
+        expect(response.setHeader.callCount).to.eq(3);
         assert(http.createServer.called);
         expect(https.createServer).to.not.have.been.called;
         assert(global.setTimeout.called);
@@ -164,6 +196,7 @@ describe('$$server', function() {
         yargs([ '--port', '9999' ]);
         $$server([ 'server', 1234 ]);
         assert(app.$$load.called);
+        expect(response.setHeader.callCount).to.eq(3);
         assert(http.createServer.called);
         expect(https.createServer).to.not.have.been.called;
         assert(global.setTimeout.called);
@@ -189,6 +222,7 @@ describe('$$server', function() {
     it('test call with https and port 443', function() {
         $$server([ 'server', 443 ]);
         assert(app.$$load.called);
+        expect(response.setHeader.callCount).to.eq(3);
         expect(http.createServer).to.not.have.been.called;
         assert(https.createServer.called);
         assert(global.setTimeout.called);
@@ -218,6 +252,7 @@ describe('$$server', function() {
         yargs([ '--usessl' ]);
         $$server([ 'server', 1234 ]);
         assert(app.$$load.called);
+        expect(response.setHeader.callCount).to.eq(3);
         expect(http.createServer).to.not.have.been.called;
         assert(https.createServer.called);
         assert(global.setTimeout.called);