+++ test/src/directives/ngie-if.spec.js
@@ -3,23 +3,16 @@ import { expect } from              'chai';
 import simple, { mock } from        'simple-mock';
 
 // System Modules
+
 import cheerio from                 'cheerio';
 
 // Angie Modules
+import $compile from                '../../../src/factories/$Compile';
+
 const TEST_ENV =                    global.TEST_ENV || 'src',
-    $$ngieIfFactory =             require(`../../../${TEST_ENV}/directives/ngie-if`);
+    $$ngieIfFactory =               require(`../../../${TEST_ENV}/directives/ngie-if`);
 
 describe('$$ngieIfFactory', function() {
-    let $,
-        el,
-        scope,
-        attrs,
-        fn;
-
-    beforeEach(function() {
-        $ = cheerio.load('<div><div class="test" ngie-if="test"></div></div>');
-        el = $('.test');
-    });
     it('test ngieIfFactory returns', function() {
         let obj = $$ngieIfFactory();
         expect(obj.priority).to.eq(1);
@@ -27,24 +20,63 @@ describe('$$ngieIfFactory', function() {
         expect(obj.link).to.be.a.function;
     });
     describe('link', function() {
-        let el,
-            scope,
-            attrs,
-            fn;
+        let template,
+            scope;
 
         beforeEach(function() {
-            let $ = cheerio.load(
-                '<div><div class="test" ngie-if="test"></div></div>'
-            );
+            template = '<div><div class="test" ngie-if="test"></div></div>';
             scope = { test: true };
-            attrs = { ngieIf: 'test' };
-            el = $('.test');
-            fn = $$ngieIfFactory().link;
         });
-        it('test ngieIf directive condition found in link, attr removed', function() {
-            fn(scope, el, attrs);
-            expect($.html()).to.eq('<div><div class="test"></div></div>');
-            expect(attrs.hasOwnProperty('ngieIf')).to.be.false;
+        it('test condition found in link, attr removed', function() {
+            $compile(template)(scope).then(function(t) {
+                expect(t).to.eq('<div><div class="test"></div></div>');
+            });
+        });
+        it('test condition found in link, attr removed, content', function() {
+            template =
+                '<div><div class="test" ngie-if="test">test</div></div>';
+            $compile(template)(scope).then(function(t) {
+                expect(t).to.eq('<div><div class="test">test</div></div>');
+            });
+        });
+        it('test condition found in link, attr removed, scope content',
+            function() {
+                template =
+                    '<div><div class="test" ngie-if="test">{{test1}}</div></div>';
+                scope.test1 = 'test';
+                $compile(template)(scope).then(function(t) {
+                    expect(t).to.eq('<div><div class="test">test</div></div>');
+                });
+            }
+        );
+        it('test condition not found, element removed', function() {
+            delete scope.test;
+            $compile(template)(scope).then(function(t) {
+                expect(t).to.eq('<div></div>');
+            });
+        });
+        it('test condition false, element removed', function() {
+            scope.test = false;
+            $compile(template)(scope).then(function(t) {
+                expect(t).to.eq('<div></div>');
+            });
+        });
+        it('test condition false, element removed, content', function() {
+            template =
+                '<div><div class="test" ngie-if="test">test</div></div>';
+            scope.test = false;
+            $compile(template)(scope).then(function(t) {
+                expect(t).to.eq('<div></div>');
+            });
+        });
+        it('test condition false, element removed, content', function() {
+            template =
+                '<div><div class="test" ngie-if="test">{{test`}}</div></div>';
+            scope.test1 = 'test';
+            scope.test = false;
+            $compile(template)(scope).then(function(t) {
+                expect(t).to.eq('<div></div>');
+            });
         });
     });
 });